refactor: relocate payments module and disable billing ui
